Remove the record button
The browser extension is now the preferred option for UI recording.
Bug fix: Now uses sanitized file names for each RSpec scenario.
