+++ lib/appmap/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module AppMap
-  VERSION = '0.15.1'
+  VERSION = '0.16.0'
 end
