--- public/appmap.css
@@ -1,52 +1,0 @@
-#appmap-record-container {
-  position: fixed;
-  bottom: 0em;
-  left: 0em;
-  width: 8em;
-  height: 2em;
-  background-color: rgba(0,0,0,0.65);
-  border-radius: .25em;
-  margin: 0.25em;
-  font-size: 16px;
-  color: #fff;
-  z-index: 99999;
-}
-
-#appmap-record-container input[type=checkbox] {
-  display: none;
-}
-
-.appmap-record-button {
-  display: inline-block;
-  width: 1em;
-  height: 1em;
-  background: red;
-  border-radius: 50%;
-  transition: border-radius 0.25s;
-  box-shadow: 0 0 0 4px #fff;
-  border: #000 solid 2px;
-  margin: 0.375em;
-  cursor: pointer;
-}
-
-.appmap-content {
-  position: inherit;
-  margin-left: 0.25em;
-  line-height: 2;
-}
-
-input[type=checkbox]:checked + .appmap-record-button {
-  transform: translate(0.05em,0.05em);
-  width: 0.85em;
-  height: 0.85em;
-  border-radius: 20%;
-}
-
-.no-select {
-  -webkit-touch-callout: none;
-  -webkit-user-select: none;
-  -khtml-user-select: none;
-  -moz-user-select: none;
-  -ms-user-select: none;
-  user-select: none;
-}
