--- public/appmap.html
@@ -1,5 +1,0 @@
-<div id="appmap-record-container">
-  <input type="checkbox" id="appmap-record" />
-  <label class="appmap-record-button" for="appmap-record" />
-  <span class="appmap-content no-select" id="appmap-record-status" />
-</div>
